Support labels for VMs
A subset of metadata applied by BOSH in the set_vm_metadata will also be
applied as labels to the VM. Labels can be viewed in the Dev Console, and
allow easy filtering and identification of VMs for operators.

The GCE labels API is in beta. This change uses the beta client library
only for the labels feature.