google: Support reboot of TERMINATED VMs
The reboot function triggered by `bosh cck` only called the `reset` API
method. That method can't be called on an instance in the TERMINATED
state (which is the state of a preemptible VM that's been preempted).
This change uses `start` to boot a terminted VM, or `reset` for a
running VM.