Add support for XPN
Surface `network_project_id` as a cloud property which we use to resolve
the path of the network to create a compute instance in.

XPN or not, the network path looks like this:
projects/[network project]/global/networks/[network name]
projects/[network projcet]/regions/[region]/subnetworks/[subnet name]

To support XPN we allow the configuration of the network project instead
of always picking the project hosting the compute instance. Default
behavior is unaffected.