ci, docs: ssh removed in new stemcells
The 3421+ stemcells can no longer be used as hosts for ssh_tunnel
meaning you can no longer SSH directly into the director but need
to use `bosh ssh`.

On GCP we recommend using a jumpbox (bosh-bastion) that has
access to the VPC with bosh resources and not exposing any of these
resources with external IPs. This jumpbox can be secured for the
environment and used as a ssh_tunnel into the cluster if
desired.

Update the docs to not use ssh_tunnel and remove the feature use in CI.

Related #199