Update BOSH and Cloud Foundry docs to use bosh-cli V2 The existing instructions have been minimally modified to use bosh-cli. https://github.com/cloudfoundry/bosh-cli
The intent of this is to be first of two commits upgrading the
docs to use GCS rather than local blobstore. GCS will require
using the new CLI, hence this change. That second half is
blocked on CPI and BOSH director changes.
cloudfoundry/bosh#1732

This includes upgrades to the latest stemcell and director.
This validates the docs for future stemcell and director changes,
which will be needed for GCS support.

Certificates were generated to be compatible with bosh-cli, which
provides no option to allow insecure connections. The new CLI
itself includes the ability to generate valid certificates.
cloudfoundry/bosh-cli#19
https://bosh.io/docs/director-certs.html

The specific command used to generate certificates was
`
bosh interpolate cert-config.yml --vars-store certs.yml -v internal_ip=10.0.0.6
`

`ca_cert.pem` is included in the docs root as it is referred to
by both the BOSH and Cloud Foundry docs. It is necessary to
login to the director.

The previous issue where enumeration wasn't happening due to
code blocks has also been corrected by adding a level
of indentation to all code blocks.