Include sane pipeline credential defaults
Currently, setting up the CI pipeline is difficult due to the
number of fields that must be configured in `credentials.yml.tpl`.
Many of those fields can be safely set to default values as they
reference GCP project-internal temporary resources.

This change introduces default values to the created resources and
networking configuration. These are known to work together and
should reduce the friction to creating the pipeline from scratch.
This reduces the number of fields to fill from 40 to 16.