action/calculate_vm_cloud_properties: implementation of experimental call
Implement the spec[1] of the new action. This implementation validates
the minimum requirements for a custom machine type[2]. It does not validate
maximum requirements because they depent on CPU architecture, region,
and avalibility. This limited validation should serve as a good starting
point but we can evaluate in the future if this needs to be expanded.

[1] http://bosh.io/docs/cpi-api-v1#calculate-vm-cloud-properties
[2] https://cloud.google.com/compute/docs/instances/creating-instance-with-custom-machine-type#specifications