google: Avoid Sandy Bridge CPUs
This change allows the CPI to specify a CPU Architecture other than
Sandy Bridge when creating a VM in a zone where it is possible to
receive that processor type.

Sandy Bridge is avoided as it does not expose RDRAND, which is required
to run rng-tools on the stemcell. rng-tools is necessary to seed entropy
to avoid crypto/rand from being blocked at boot, which may cause the
director to timeout waiting for the bosh-agent to become available.