Make unit tests work again
This change makes it possible to inject a Google client when we're testing.
And then it fixes all the broken tests.

To support CPI config, we can't create the Google API client as early as
we did previously (from the base BOSH config). Because we're creating
the client later in the lifecycle of a CPI request, it messed up
testing.

This new method has the benefit of exercising some of the CPI request
context stuff.