Create stemcell images with MULTI_IP_SUBNET guest os feature
This change updates create_stemcell to use the MULTI_IP_SUBNET guest os
feature when creating the google cloud image. This allows the instance
to use a different netmask then the /32 that is enforced by google
cloud. See
https://cloud.google.com/vpc/docs/create-use-multiple-interfaces for
more details.

This improves the resilience of network diffing for the BOSH director
and prevents unnecessary recreates. Specifically, there are times in
which the BOSH director will save the network settings as reported by
the agent in the database. Due to the fact that the agent on google
cloud reports a /32 netmask, the BOSH director will detect a network change on the
next deploy when it computes the expected netmask from the manual
network's CIDR. This will cause a recreation of the VM. With this
change, this should no longer happen.

[#166359647](https://www.pivotaltracker.com/story/show/166359647)