Stop printing the contents of the Google SAK
Under certain circumstances, the CPI would print the contents of
`cpi.json` in the DEBUG logs for the operation. This would be fine,
but that file contains the GCP Service Account Key!

This commit adds the "Quiet" flag to the file read, which suppresses
the code that prints the contents of `cpi.json`.

[#176466721] CPIs are leaking iaas credentials in bosh debug output

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>