Fix for uploading heavy stemcells
It seems google has changed the API. Previously it was fine to create an
image from a raw disk image in GCS using only a client that had the compute scope.

This started failing with the error "Required 'read' permission" even though
the GCS file was public. Using a client that has both compute and storage
scopes at the same time seems to resolve the problem.