Add sole-tenancy instance to clean up list
The sole-tenancy instances are quite expensive, that is why we should
make sure that they are not leaked.

There is a test which deletes the sole-tenancy instance but
in case something fails the instance should be deleted in the clean up
loop of the test suite. The "integration-delete" tag is set for that